+++ tests/unit/test_s3_manager.py
@@ -42,42 +42,52 @@ class TestS3Manager:
         with pytest.raises(S3Error):
             testInstance.createS3Bucket('testBucket', 'test-perms')
 
-    def test_putObjectInBucket_success(self, testInstance, mocker):
-        mockHash = mocker.patch.object(S3Manager, 'getmd5HashOfObject')
-        mockHash.return_value = 'testMd5Hash'
-        mockGet = mocker.patch.object(S3Manager, 'getObjectFromBucket')
-        mockGet.return_value = None
+    def test_putObjectInBucket_success_epub(self, testInstance, mocker):
+        managerMocks = mocker.patch.multiple(
+            S3Manager,
+            getmd5HashOfObject=mocker.DEFAULT,
+            getObjectFromBucket=mocker.DEFAULT,
+            putExplodedEpubComponentsInBucket=mocker.DEFAULT
+        )
+        managerMocks['getmd5HashOfObject'].return_value = 'testMd5Hash'
+        managerMocks['getObjectFromBucket'].return_value = None
 
         testInstance.s3Client.put_object.return_value = 'testObjectResponse'
 
-        testResponse = testInstance.putObjectInBucket('testObj', 'testKey', 'testBucket')
+        testResponse = testInstance.putObjectInBucket('testObj', 'testKey.epub', 'testBucket')
 
         assert testResponse == 'testObjectResponse'
 
-        mockHash.assert_called_once_with('testObj')
-        mockGet.assert_called_once_with('testKey', 'testBucket', md5Hash='testMd5Hash')
+        managerMocks['getmd5HashOfObject'].assert_called_once_with('testObj')
+        managerMocks['getObjectFromBucket'].assert_called_once_with('testKey.epub', 'testBucket', md5Hash='testMd5Hash')
+        managerMocks['putExplodedEpubComponentsInBucket'].assert_called_once_with('testObj', 'testKey.epub', 'testBucket') 
         testInstance.s3Client.put_object.assert_called_once_with(
-            ACL='public-read', Body='testObj', Bucket='testBucket', Key='testKey'
+            ACL='public-read', Body='testObj', Bucket='testBucket', Key='testKey.epub'
         )
 
     def test_putObjectInBucket_existing_outdated(self, testInstance, mocker):
-        mockHash = mocker.patch.object(S3Manager, 'getmd5HashOfObject')
-        mockHash.return_value = 'testMd5Hash'
         mockExisting = mocker.MagicMock()
         mockExisting.statusCode = 301
-        mockGet = mocker.patch.object(S3Manager, 'getObjectFromBucket')
-        mockGet.return_value = mockExisting
+        managerMocks = mocker.patch.multiple(
+            S3Manager,
+            getmd5HashOfObject=mocker.DEFAULT,
+            getObjectFromBucket=mocker.DEFAULT,
+            putExplodedEpubComponentsInBucket=mocker.DEFAULT
+        )
+        managerMocks['getmd5HashOfObject'].return_value = 'testMd5Hash'
+        managerMocks['getObjectFromBucket'].return_value = mockExisting
 
         testInstance.s3Client.put_object.return_value = 'testObjectResponse'
 
-        testResponse = testInstance.putObjectInBucket('testObj', 'testKey', 'testBucket')
+        testResponse = testInstance.putObjectInBucket('testObj', 'testKey.epub', 'testBucket')
 
         assert testResponse == 'testObjectResponse'
 
-        mockHash.assert_called_once_with('testObj')
-        mockGet.assert_called_once_with('testKey', 'testBucket', md5Hash='testMd5Hash')
+        managerMocks['getmd5HashOfObject'].assert_called_once_with('testObj')
+        managerMocks['getObjectFromBucket'].assert_called_once_with('testKey.epub', 'testBucket', md5Hash='testMd5Hash')
+        managerMocks['putExplodedEpubComponentsInBucket'].assert_called_once_with('testObj', 'testKey.epub', 'testBucket') 
         testInstance.s3Client.put_object.assert_called_once_with(
-            ACL='public-read', Body='testObj', Bucket='testBucket', Key='testKey'
+            ACL='public-read', Body='testObj', Bucket='testBucket', Key='testKey.epub'
         )
 
     def test_putObjectInBucket_existing_unmodified(self, testInstance, mocker):
@@ -86,12 +96,12 @@ class TestS3Manager:
         mockGet = mocker.patch.object(S3Manager, 'getObjectFromBucket')
         mockGet.return_value = {'ResponseMetadata': {'HTTPStatusCode': 304}}
 
-        testResponse = testInstance.putObjectInBucket('testObj', 'testKey', 'testBucket')
+        testResponse = testInstance.putObjectInBucket('testObj', 'testKey.epub', 'testBucket')
 
         assert testResponse == None
 
         mockHash.assert_called_once_with('testObj')
-        mockGet.assert_called_once_with('testKey', 'testBucket', md5Hash='testMd5Hash')
+        mockGet.assert_called_once_with('testKey.epub', 'testBucket', md5Hash='testMd5Hash')
         testInstance.s3Client.put_object.assert_not_called
 
     def test_putObjectInBucket_error(self, testInstance, mocker):
@@ -104,6 +114,26 @@ class TestS3Manager:
 
         with pytest.raises(S3Error):
             testInstance.putObjectInBucket('testObj', 'testKey', 'testBucket')
+
+    def test_putExplodedEpubComponentsInBucket(self, testInstance, mocker):
+        mockZipComp = mocker.MagicMock()
+        mockZipComp.read.side_effect = ['compBytes1', 'compBytes2']
+
+        mockZip = mocker.MagicMock()
+        mockZip.namelist.return_value = ['comp1', 'comp2']
+        mockZip.open.return_value = mockZipComp
+
+        mockZipFile = mocker.patch('managers.s3.ZipFile')
+        mockZipFile.return_value.__enter__.return_value = mockZip
+
+        mockPut = mocker.patch.object(S3Manager, 'putObjectInBucket')
+
+        testInstance.putExplodedEpubComponentsInBucket(b'testObj', 'testKey.epub', 'testBucket')
+
+        mockPut.assert_has_calls([
+            mocker.call('compBytes1', 'testkey/comp1', 'testBucket'),
+            mocker.call('compBytes2', 'testkey/comp2', 'testBucket')
+        ])
 
     def test_getObjectFromBucket_success(self, testInstance, mocker):
         testInstance.s3Client.get_object.return_value = 'testObject'